+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Target Scheduler
 
+## 3.3.0.0 - 2023-08-XX
+* Added ability to override exposure ordering.
+
 ## 3.2.1.0 - 2023-08-09
 * Fixed bug preventing target ROI from being applied properly.